pyproject.toml
[project]
name = "github-issue-tracker"
version = "0.1.0"
description = "Interactive GitHub issue tracker with YAML templates"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "textual>=0.79.0",
    "rich>=13.7.0",
    "click>=8.1.7",
    "pydantic>=2.8.0",
    "python-dateutil>=2.9.0",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.12.0",
    "structlog>=23.1.0",
]

[project.scripts]
gh-tracker = "github_issue_tracker.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true